Fix cargo clippy
