--- crates/runtime/src/datafusion/functions/geospatial/measurement/mod.rs
@@ -1,30 +1,0 @@
-// Licensed to the Apache Software Foundation (ASF) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The ASF licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
-mod area;
-mod contains;
-mod distance;
-mod within;
-
-use datafusion::prelude::SessionContext;
-
-pub fn register_udfs(ctx: &SessionContext) {
-    ctx.register_udf(area::Area::new().into());
-    ctx.register_udf(distance::Distance::new().into());
-    ctx.register_udf(contains::Contains::new().into());
-    ctx.register_udf(within::Within::new().into());
-}
