--- crates/runtime/src/datafusion/functions/geospatial/mod.rs
@@ -1,30 +1,0 @@
-// Licensed to the Apache Software Foundation (ASF) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The ASF licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
-pub mod accessors;
-pub mod constructors;
-pub mod data_types;
-pub mod error;
-mod measurement;
-
-use datafusion::prelude::SessionContext;
-
-pub fn register_udfs(ctx: &SessionContext) {
-    constructors::register_udfs(ctx);
-    accessors::register_udfs(ctx);
-    measurement::register_udfs(ctx);
-}
